requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dateparser==1.0.0
decorator==4.4.2
gevent==21.1.2
greenlet==1.0.0
idna==2.10
ipykernel==5.5.0
jedi==0.18.0
kiwisolver==1.3.1
matplotlib==3.3.4
memory-profiler==0.58.0
numpy>=1.16.1
pandas>=1.1.2
pandas-ta==0.2.23b0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
prompt-toolkit==3.0.16
psutil==5.8.0
ptyprocess==0.7.0
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
wcwidth==0.2.5
websockets>=8.1
yarg==0.1.9
zope.event==4.5.0
zope.interface==5.2.0
